Kang api

Raytracing Functions

Kang.AbstractMetric — Type

abstract type AbstractMetric

Abstract Metric Type

Kang.Kerr — Type

Kerr Metric in Boyer Lindquist Coordinates

Kang.Gθ — Method

Gθ(metric::Kerr{T}, α, β, θs, θo, isindir, n) where T

Returns the antiderivative $G_\theta=\int\frac{d\theta}{\sqrt{\Theta(\theta)}}$

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image ordered by minotime

Kang.Ir — Method

Ir(metric::Kerr{T}, νr::Bool, θo, rs, α, β) where T

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$

metric: Kerr{T} metric

νr : Radial velocity direction at emission

θo : Observer inclination

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

Kang._isreal2 — Method

_isreal2(num, T)

Checks if a complex number is real to within machine precision

`num`: Complex number

`T`  : Type of number

Kang.emission_coordinates — Method

emission_coordinates(metric::Kerr{T}, α, β, θs, θo, isindir, n) where T

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β) for an observer located at inclination θo.

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.emission_coordinates_fast_light — Method

emissioncoordinatesfast_light(metric::Kerr{T}, α, β, θs, θo, isindir, n) where T

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β) for an observer located at inclination θo.

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.emission_inclination — Method

emission_inclination(metric::Kerr{T}, α, β, θo, τ) where T

Emission inclination for point that appears at the screen coordinate (α, β) for an observer located at inclination θo that is connected by a ray with mino time τ.

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.emission_radius — Method

emission_radius(metric::Kerr{T}, α, β, θs, θo, isindir, n) where T

Emission radius for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β) for an observer located at inclination θo.

metric : Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Normalized angular momentum

n : Image index

Kang.get_radial_roots — Method

getradialroots(metric::Kerr{T}, η, λ) where T

Returns roots of $r^4 + (a^2-η-λ^2)r^2 + 2(η+(a-λ)^2)r - a^2η$

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal agular momentum

Kang.horizon — Method

horizon(metric::Kerr{T})

Params

metric : Kerr{T} metric

Outer Horizon for the Kerr{T} metric.

Kang.metric_dd — Method

metric_dd(metric::AbstractMetric, args)

Returns the metric in some representation (usually as an nxn matrix).

Kang.metric_dd — Method

metric_dd(metric::Kerr{T}, r, θ)

Params

metric : Kerr{T} metric

r : Radius

θ : Inclination

Inverse Kerr{T} metric in Boyer Lindquist (BL) coordinates.

Kang.metric_dd — Method

metric_dd(metric::Kerr{T}, coordinates)

Params

metric : Kerr{T} metric

coordinates : Coordinates (t, r, θ, ϕ)

Kerr{T} metric in Boyer Lindquist (BL) coordinates.

Kang.metric_uu — Method

metric_uu(metric::AbstractMetric, args)

Returns the inverse metric in some representation (usually as an nxn matrix).

Kang.metric_uu — Method

metric_uu(metric::Kerr{T}, r, θ)

Params

metric : Kerr{T} metric

r : Radius

θ : Inclination

Inverse Kerr{T} metric in Boyer Lindquist (BL) coordinates.

Kang.metric_uu — Method

metric_uu(metric::Kerr{T}, coordinates)

Params

metric : Kerr{T} metric

coordinates : Coordinates (t, r, θ, ϕ)

Inverse Kerr{T} metric in Boyer Lindquist (BL) coordinates.

Kang.mino_time — Method

mino_time(η, λ, a, θs, θo, isindir::Bool, n::Int64)

Mino time of trajectory between two inclinations for a given screen coordinate

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

a : Blackhole angular Momentum

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image in orde of amount of minotime traversed

Kang.r_potential — Method

r_potential(metric::Kerr{T}, r, η, λ) where T

Radial potential of spacetime

`metric`: Kerr{T} metric

`η`  : Reduced Carter constant

`λ`  : Reduced azimuthal agular momentum

`r`  : Boyer Lindquist radius

Kang.radial_integrals_case2 — Method

radial_integrals_case2(metric::Kerr{T}, rs, roots::NTuple{4}, τ, νr)

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Kang.radial_integrals_case3 — Method

radial_integrals_case3(metric::Kerr{T}, rs, roots::NTuple{4}, τ) where {T}

Returns the radial integrals for the case where there are two real roots in the radial potential

Kang.radial_integrals_case4 — Method

radial_integrals_case4(metric::Kerr{T}, rs, roots::NTuple{4}, τ)

Returns the radial integrals for the case where there are no real roots in the radial potential

Kang.regularized_Pi — Method

regularized_Pi(n, ϕ, k)

Regularized elliptic integral of the third kind

    `n`: Parameter

    `ϕ`: Argument

    `k`: Parameter

Kang.α — Method

α(metric::Kerr{T}, α, θo) where T

Horizontal Bardeen Screen Coordinate metric: Kerr

`α`: Horizontal Bardeen screen coordinate

`θo`: Observer inclination

Kang.αboundary — Method

αboundary(metric::Kerr, θs) where

Defines a horizontal boundary on the assmyptotic observers screen where emission that originates from θs must fall within.

`metric`: Kerr metric

`θs`  : Emission Inclination

Kang.β — Method

β(metric::Kerr, λ, η, θo)

Horizontal Bardeen Screen Coordinate metric: Kerr

`λ`: Energy reduced Azimuthal angular momentul

`η`: Energy reduced Carter integral 

`θo`: Observer inclination

Kang.βboundary — Method

βboundary(metric::Kerr{T}, α, θo, θs) where {T}

Defines a vertical boundary on the Assyptotic observers screen where emission that originates from θs must fall within.

`metric`: Kerr{T} metric

`α`   : Horizontal Bardeen screen coordinate

`θo`  : Observer inclination

`θs`  : Emission Inclination

Kang.η — Method

η(metric::Kerr{T}, α, β, θo) where T

Energy reduced Carter integral metric: Kerr

`α`: Horizontal Bardeen screen coordinate

`β`: Bardeen vertical coordinate

`θo`: Observer inclination

Kang.θ_potential — Method

θ_potential(metric::Kerr{T}, r, η, λ) where T

Theta potential of a kerr blackhole

`metric`: Kerr{T} metric

`η`  : Reduced Carter constant

`λ`  : Reduced azimuthal agular momentum

`θ`  : Boyer Lindquist inclination

Kang.λ — Method

λ(metric::Kerr{T}, α, θo) where T

Energy reduced azimuthal angular momentum metric: Kerr

`α`: Horizontal Bardeen screen coordinate

`θo`: Observer inclination