Kang api

Kang.AbstractMetric

Kang.Kerr

Kang.Gθ

Kang.Ir

Kang.Ir

Kang._isreal2

Kang.calcPol

Kang.emission_coordinates

Kang.emission_coordinates_fast_light

Kang.emission_inclination

Kang.emission_inclination

Kang.emission_radius

Kang.emission_radius

Kang.get_radial_roots

Kang.horizon

Kang.jac_bl_d_zamo_u

Kang.jac_bl_u_zamo_d

Kang.jac_fluid_u_zamo_d

Kang.jac_zamo_d_bl_u

Kang.jac_zamo_u_bl_d

Kang.metric_dd

Kang.metric_uu

Kang.mino_time

Kang.p_bl_d

Kang.penrose_walker

Kang.r_potential

Kang.radial_integrals_case2

Kang.radial_integrals_case3

Kang.radial_integrals_case4

Kang.raytrace

Kang.regularized_Pi

Kang.screen_polarisation

Kang.α

Kang.αboundary

Kang.β

Kang.βboundary

Kang.η

Kang.θ_potential

Kang.λ

Raytracing Functions

Kang.emission_radius — Method

emission_radius(
    metric::Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> Tuple{Any, Bool, Bool}

Emission radius for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Is emission to observer direct or indirect

n : Image index

Kang.emission_radius — Method

emission_radius(
    metric::Kerr{T},
    α,
    β,
    τ,
    θo
) -> Tuple{Any, Any, Any}

Emission radius for point originating at at Mino time τ whose image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

-metric : Kerr{T} metric -α : Horizontal Bardeen screen coordinate -β : Vertical Bardeen screen coordinate -θs : Emission inclination -θo : Observer inclination -isindir : Normalized angular momentum -n : Image index

Kang.emission_inclination — Method

emission_inclination(
    metric::Kerr{T},
    α,
    β,
    θo,
    rs,
    νr
) -> NTuple{6, Any}

Emission inclination for point originating at inclination rs whose nth order image appears at screen coordinate (α, β).

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.emission_inclination — Method

emission_inclination(
    metric::Kerr{T},
    α,
    β,
    τ,
    θo
) -> NTuple{6, Any}

Emission inclination for point at Mino time τ whose image appears at screen coordinate (α, β).

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.emission_coordinates_fast_light — Function

emission_coordinates_fast_light(
    metric::Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> Tuple{Any, Any, Any, Bool, Bool}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.emission_coordinates — Function

emission_coordinates(
    metric::Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> NTuple{6, Any}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.raytrace — Function

raytrace(metric::Kerr{T}, α, β, θo, τ) -> NTuple{6, Any}

Raytrace a point that appears at the screen coordinate (α, β) for an observer located at inclination θo

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.mino_time — Function

mino_time(metric::Kerr{T}, α, β, θs, θo, isindir, n) -> Any

Mino time of trajectory between two inclinations for a given screen coordinate

Arguments

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

a : Blackhole angular Momentum

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image in orde of amount of minotime traversed

Metric Functions

Kang.AbstractMetric — Type

abstract type AbstractMetric

Abstract Metric Type

Kang.Kerr — Type

struct Kerr{T} <: Kang.AbstractMetric

Kerr Metric in Boyer Lindquist Coordinates

mass: M = mass

spin: a = J/M, where J is the angular momentum and M is the mass of the blackhole.

Kang.metric_uu — Function

metric_uu(metric::Kang.AbstractMetric, args...) -> Any

Returns the inverse metric in some representation (usually as an nxn matrix).

metric_uu(metric::Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

metric_uu(metric::Kerr{T}, coordinates) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Kang.metric_dd — Function

metric_dd(metric::Kang.AbstractMetric, args...) -> Any

Returns the metric in some representation (usually as an nxn matrix).

metric_dd(metric::Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

metric_dd(metric::Kerr{T}, coordinates) -> Any

Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Kang.horizon — Function

horizon(metric::Kerr{T}) -> Any

Outer Horizon for the Kerr metric.

Kang.λ — Function

λ(_::Kerr, α, θo) -> Any

Energy reduced azimuthal angular momentum

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination

Kang.η — Function

η(metric::Kerr, α, β, θo) -> Any

Energy reduced Carter integral

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

β: Bardeen vertical coordinate

θo: Observer inclination

Radial Integrals

Kang.r_potential — Function

r_potential(metric::Kerr{T}, η, λ, r) -> Any

Radial potential of spacetime

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

r : Boyer Lindquist radius

Kang.get_radial_roots — Function

get_radial_roots(metric::Kerr{T}, η, λ) -> NTuple{4, Any}

Returns roots of $r^4 + (a^2-η-λ^2)r^2 + 2(η+(a-λ)^2)r - a^2η$

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

Kang.Ir — Method

Ir(metric::Kerr, νr::Bool, θo, rs, α, β) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

θo : Observer inclination

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

Kang.Ir — Method

Ir(metric::Kerr{T}, νr::Bool, rs, η, λ) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

θo : Observer inclination

η : Reduced Carter Constant

λ : Reduced Azimuthal Angular Momentum

Kang.radial_integrals_case2 — Function

radial_integrals_case2(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ,
    νr
) -> NTuple{5, Any}

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Kang.radial_integrals_case3 — Function

radial_integrals_case3(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> NTuple{5, Any}

Returns the radial integrals for the case where there are two real roots in the radial potential

Kang.radial_integrals_case4 — Function

radial_integrals_case4(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> Tuple

Returns the radial integrals for the case where there are no real roots in the radial potential

Angular Integrals

Kang.θ_potential — Function

θ_potential(metric::Kerr{T}, η, λ, θ) -> Any

Theta potential of a Kerr blackhole

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

θ : Boyer Lindquist inclination

Kang.Gθ — Function

Gθ(
    metric::Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> NTuple{5, Any}

Returns the antiderivative $G_\theta=\int\frac{d\theta}{\sqrt{\Theta(\theta)}}$. See θ_potential(x) for an implementation of $\Theta( heta)$.

Arguments

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image ordered by minotime

Screen Coordinates

Kang.α — Function

α(_::Kerr, λ, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination

Kang.β — Function

β(metric::Kerr, λ, η, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

λ: Energy reduced Azimuthal angular momentul

η: Energy reduced Carter integral

θo: Observer inclination

Kang.αboundary — Function

αboundary(metric::Kerr, θs) -> Any

Defines a horizontal boundary on the assmyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr metric

θs : Emission Inclination

Kang.βboundary — Function

βboundary(metric::Kerr{T}, α, θo, θs) -> Any

Defines a vertical boundary on the Assyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

θo : Observer inclination

θs : Emission Inclination

Polarization Functions

Kang.p_bl_d — Function

p_bl_d(metric::Kerr{T}, r, θ, η, λ, νr, νθ) -> Any

Returns the momentum form in the Boyer-Lindquist basis.

Kang.jac_bl_u_zamo_d — Function

jac_bl_u_zamo_d(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts ZAMO vector to a Boyer-Lindquist basis

Kang.jac_zamo_u_bl_d — Function

jac_zamo_u_bl_d(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts Boyer-Lindquist vector to a ZAMO basis

Kang.jac_bl_d_zamo_u — Function

jac_bl_d_zamo_u(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts ZAMO covector to a Boyer-Lindquist basis

Kang.jac_zamo_d_bl_u — Function

jac_zamo_d_bl_u(metric::Kerr{T}, r, θ) -> Any

Returns the Jacobian which converts a Boyer-Lindquist covector to ZAMO basis.

Kang.jac_fluid_u_zamo_d — Function

jac_fluid_u_zamo_d(_::Kerr{T}, β, θ, φ) -> Any

Jacobian which expreases ZAMO vector in the fluid frame

Kang.screen_polarisation — Function

screen_polarisation(
    metric::Kerr{T},
    κ::Complex,
    θ,
    α,
    β
) -> Tuple{Any, Any}

Returns the screen polarization associated with a killing spinor κ as seen seen by an assymptotic observer.

Kang.penrose_walker — Function

penrose_walker(
    metric::Kerr{T},
    r,
    θ,
    p_u::AbstractVector,
    f_u::AbstractVector
) -> Tuple{Any, Any}

Returns the Penrose walker constant for a photon with momentum pu emitted from a fluid particle with momentum fu.

Kang.calcPol — Function

calcPol(
    metric::Kerr{T},
    α,
    β,
    ri,
    θs,
    θo,
    magfield::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    βfluid::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    νr::Bool,
    θsign::Bool
) -> NTuple{4, Any}

Calculates the polarization of a photon emitted from a fluid particle with momentum f_u and observed by an asymptotic observer.

Misc

Kang._isreal2 — Function

_isreal2(num::Complex{T}) -> Any

Checks if a complex number is real to some tolerance

Kang.regularized_Pi — Function

regularized_Pi(n, ϕ, k) -> Any

Regularized elliptic integral of the third kind

Arguments

n: Parameter

ϕ: Arguments

k: Parameter