Krang api

Krang.AbstractCamera

Krang.AbstractGeometry

Krang.AbstractMaterial

Krang.AbstractMetric

Krang.AbstractPixel

Krang.AbstractScreen

Krang.ConeGeometry

Krang.ElectronSynchrotronPowerLawPolarization

Krang.IntensityCamera

Krang.IntensityPixel

Krang.IntensityScreen

Krang.Kerr

Krang.Mesh

Krang.SlowLightIntensityCamera

Krang.SlowLightIntensityPixel

Krang.SlowLightIntensityScreen

Krang.UnionGeometry

Krang.Gθ

Krang.Ir

Krang.Ir_inf

Krang.Ir_s

Krang.It

Krang.It_inf

Krang.It_m_I0_terms

Krang.It_w_I0_terms

Krang.Iϕ

Krang.Iϕ_inf

Krang.Iϕ_m_I0_terms

Krang.Iϕ_w_I0_terms

Krang._isreal2

Krang.emission_azimuth

Krang.emission_coordinates

Krang.emission_coordinates_fast_light

Krang.emission_inclination

Krang.emission_radius

Krang.get_radial_roots

Krang.horizon

Krang.jac_bl_d_zamo_u

Krang.jac_bl_u_zamo_d

Krang.jac_fluid_u_zamo_d

Krang.jac_zamo_d_bl_u

Krang.jac_zamo_u_bl_d

Krang.metric_dd

Krang.metric_uu

Krang.mino_time

Krang.p_bl_d

Krang.penrose_walker

Krang.r_potential

Krang.radial_inf_integrals_case2

Krang.radial_inf_integrals_case3

Krang.radial_inf_integrals_case4

Krang.radial_integrals

Krang.radial_m_I0_terms_integrals_case2

Krang.radial_m_I0_terms_integrals_case3

Krang.radial_m_I0_terms_integrals_case4

Krang.radial_w_I0_terms_integrals_case2

Krang.radial_w_I0_terms_integrals_case3

Krang.radial_w_I0_terms_integrals_case4

Krang.raytrace

Krang.regularized_Pi

Krang.screen_polarisation

Krang.synchrotronIntensity

Krang.synchrotronPolarization

Krang.α

Krang.αboundary

Krang.β

Krang.βboundary

Krang.η

Krang.θ_potential

Krang.λ

Raytracing Functions

Krang.emission_radius — Function

emission_radius(
    pix::Krang.AbstractPixel,
    θs,
    isindir,
    n
) -> Tuple

Emission radius for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

pix : Pixel information

θs : Emission inclination

isindir : Is emission to observer direct or indirect

n : Image index

emission_radius(
    pix::Krang.AbstractPixel,
    τ
) -> Tuple{Any, Any, Any}

Emission radius for point originating at at Mino time τ whose image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

-pix : Pixel information

τ : Mino time

Krang.emission_inclination — Function

emission_inclination(pix::Krang.AbstractPixel, rs, νr)

Emission inclination for point originating at inclination rs whose nth order image appears at screen coordinate (α, β).

Arguments

pix : Pixel information

rs : Emission radius

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

emission_inclination(
    pix::Krang.AbstractPixel,
    τ
) -> NTuple{6, Any}

Emission inclination for point at Mino time τ whose image appears at screen coordinate (α, β).

Arguments

pix : Pixel information

τ : Mino Time

Krang.emission_coordinates_fast_light — Function

emission_coordinates_fast_light(
    pix::Krang.AbstractPixel,
    θs,
    isindir,
    n
) -> NTuple{5, Any}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

pix : Pixel information

θs : Emission Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Krang.emission_coordinates — Function

emission_coordinates(
    pix::Krang.AbstractPixel,
    θs,
    isindir,
    n
) -> NTuple{6, Any}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

pix : Pixel information

θs : Emission Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Krang.raytrace — Function

raytrace(pix::Krang.AbstractPixel, τ) -> NTuple{6, Any}

Raytrace a point that appears at the screen coordinate (α, β) for an observer located at inclination θo

Arguments

pix : Pixel information

τ : Mino Time

Krang.mino_time — Function

mino_time(pix, rs, isindir) -> Any

Mino time of trajectory between an observer at infinity and point at radius rs

Arguments

pix : Pixel information

rs : Emission radius

isindir : Is the path direct or indirect?

mino_time(pix::Krang.AbstractPixel, θs, isindir, n) -> Any

Mino time of trajectory between two inclinations for a given screen coordinate

Arguments

pix : Pixel information

θs : Emission inclination

isindir : Is the path direct or indirect?

n : nth image in orde of amount of minotime traversed

Metric Functions

Krang.AbstractMetric — Type

abstract type AbstractMetric

Abstract Metric Type

Krang.Kerr — Type

struct Kerr{T} <: Krang.AbstractMetric

Kerr Metric in Boyer Lindquist Coordinates

mass: M = mass

spin: a = J/M, where J is the angular momentum and M is the mass of the blackhole.

Krang.metric_uu — Function

metric_uu(metric::Krang.AbstractMetric, args...) -> Any

Returns the inverse metric in some representation (usually as an nxn matrix).

metric_uu(metric::Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

metric_uu(metric::Kerr{T}, coordinates) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Krang.metric_dd — Function

metric_dd(metric::Krang.AbstractMetric, args...) -> Any

Returns the metric in some representation (usually as an nxn matrix).

metric_dd(metric::Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

metric_dd(metric::Kerr{T}, coordinates) -> Any

Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Krang.horizon — Function

horizon(metric::Kerr{T}) -> Any

Outer Horizon for the Kerr metric.

Krang.λ — Function

λ(_::Kerr, α, θo) -> Any

Energy reduced azimuthal angular momentum

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination

Krang.η — Function

η(metric::Kerr, α, β, θo) -> Any

Energy reduced Carter integral

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

β: Bardeen vertical coordinate

θo: Observer inclination

Radial Integrals

Krang.r_potential — Function

r_potential(metric::Kerr{T}, η, λ, r) -> Any

Radial potential of spacetime

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

r : Boyer Lindquist radius

Krang.get_radial_roots — Function

get_radial_roots(metric::Kerr{T}, η, λ) -> NTuple{4, Any}

Returns roots of $r^4 + (a^2-η-λ^2)r^2 + 2(η+(a-λ)^2)r - a^2η$

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

Krang.Ir — Function

Ir(pix::Krang.AbstractPixel, νr::Bool, rs) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

pix : Pixel information

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

rs : Emission radius

Angular Integrals

Krang.θ_potential — Function

θ_potential(metric::Kerr{T}, η, λ, θ) -> Any

Theta potential of a Kerr blackhole

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal angular momentum

θ : Boyer Lindquist inclination

Krang.Gθ — Function

Gθ(
    pix::Krang.AbstractPixel,
    θs,
    isindir,
    n
) -> NTuple{5, Any}

Returns the antiderivative $G_\theta=\int\frac{d\theta}{\sqrt{\Theta(\theta)}}$. See θ_potential(x) for an implementation of $\Theta( heta)$.

Arguments

metric: Kerr{T} metric

pix : Pixel information

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image ordered by minotime

Screen Coordinates

Krang.α — Function

α(_::Kerr, λ, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination

Krang.β — Function

β(metric::Kerr, λ, η, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

λ: Energy reduced Azimuthal angular momentul

η: Energy reduced Carter integral

θo: Observer inclination

Krang.αboundary — Function

αboundary(metric::Kerr, θs) -> Any

Defines a horizontal boundary on the assmyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr metric

θs : Emission Inclination

Krang.βboundary — Function

βboundary(metric::Kerr{T}, α, θo, θs) -> Any

Defines a vertical boundary on the Assyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

θo : Observer inclination

θs : Emission Inclination

Polarization Functions

Krang.p_bl_d — Function

p_bl_d(
    metric::Kerr{T},
    r,
    θ,
    η,
    λ,
    νr::Bool,
    νθ::Bool
) -> Any

Returns the momentum form in the Boyer-Lindquist basis.

Krang.jac_bl_u_zamo_d — Function

jac_bl_u_zamo_d(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts ZAMO vector to a Boyer-Lindquist basis

Krang.jac_zamo_u_bl_d — Function

jac_zamo_u_bl_d(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts Boyer-Lindquist vector to a ZAMO basis

Krang.jac_bl_d_zamo_u — Function

jac_bl_d_zamo_u(metric::Kerr{T}, r, θ) -> Any

Jacobian which converts ZAMO covector to a Boyer-Lindquist basis

Krang.jac_zamo_d_bl_u — Function

jac_zamo_d_bl_u(metric::Kerr{T}, r, θ) -> Any

Returns the Jacobian which converts a Boyer-Lindquist covector to ZAMO basis.

Krang.jac_fluid_u_zamo_d — Function

jac_fluid_u_zamo_d(_::Kerr{T}, β, θ, φ) -> Any

Jacobian which expreases ZAMO vector in the fluid frame

Krang.screen_polarisation — Function

screen_polarisation(
    metric::Kerr{T},
    κ::Complex,
    θ,
    α,
    β
) -> Tuple{Any, Any}

Returns the screen polarization associated with a killing spinor κ as seen seen by an assymptotic observer.

Krang.penrose_walker — Function

penrose_walker(
    metric::Kerr{T},
    r,
    θ,
    p_u::AbstractVector,
    f_u::AbstractVector
) -> Tuple{Any, Any}

Returns the Penrose walker constant for a photon with momentum pu emitted from a fluid particle with momentum fu.

Krang.synchrotronIntensity — Function

synchrotronIntensity(
    metric::Kerr{T},
    α,
    β,
    ri,
    θs,
    θo,
    magfield::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    βfluid::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    νr::Bool,
    θsign::Bool
) -> Tuple{Any, Any, Any}

Calculates the intensity of a photon emitted from a fluid particle with momentum f_u and observed by an asymptotic observer.

Krang.synchrotronPolarization — Function

synchrotronPolarization(
    metric::Kerr{T},
    α,
    β,
    ri,
    θs,
    θo,
    magfield::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    βfluid::StaticArraysCore.SArray{Tuple{3}, T, 1, 3},
    νr::Bool,
    θsign::Bool
) -> NTuple{4, Any}

Calculates the polarization of a photon emitted from a fluid particle with momentum f_u and observed by an asymptotic observer.

Raytracing API Related Functions

Krang.Mesh — Type

struct Mesh{G<:Krang.AbstractGeometry, M<:Krang.AbstractMaterial}

Krang.ElectronSynchrotronPowerLawPolarization — Type

struct ElectronSynchrotronPowerLawPolarization <: Krang.AbstractMaterial

Linear polarization material from https://doi.org/10.3847/1538-4357/abf117

Krang.UnionGeometry — Type

struct UnionGeometry{G1, G2} <: Krang.AbstractGeometry

Geometry that is comprised of the union of two geometries.

Misc

Krang._isreal2 — Function

_isreal2(num::Complex{T}) -> Any

Checks if a complex number is real to some tolerance

Krang.regularized_Pi — Function

regularized_Pi(n, ϕ, k) -> Any

Regularized elliptic integral of the third kind

Arguments

n: Parameter

ϕ: Arguments

k: Parameter

Krang.radial_m_I0_terms_integrals_case3 — Function

radial_m_I0_terms_integrals_case3(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T
) -> Tuple

Returns the radial integrals for the case where there are two real roots in the radial potential

Krang.radial_inf_integrals_case4 — Function

radial_inf_integrals_case4(
    metric::Kerr{T},
    roots::NTuple{4, T} where T
) -> NTuple{4, Any}

Returns the radial integrals for the case where there are no real roots in the radial potential

Krang.It_m_I0_terms — Function

It_m_I0_terms(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    λ,
    νr
) -> Any

Returns the antiderivative $I_t=\int\frac{r^2\Delta+2Mr(r^2+a^2-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ without I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

roots : Radial roots

λ : Reduced azimuthal angular momentum

Krang.It_w_I0_terms — Function

It_w_I0_terms(
    metric::Kerr{T},
    rs,
    τ,
    roots::NTuple{4, T} where T,
    λ,
    νr
) -> Any

Returns the antiderivative $I_t=\int\frac{r^2\Delta+2Mr(r^2+a^2-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ with I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

roots : Radial roots

λ : Reduced azimuthal angular momentum

Krang.Ir_inf — Function

Ir_inf(metric::Kerr{T}, roots) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$ evaluated at infinity. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

roots : Roots of the radial potential

Krang.emission_azimuth — Function

emission_azimuth(
    pix::Krang.AbstractPixel,
    θs,
    rs,
    τ,
    νr,
    isindir,
    n
) -> Any

Emission azimuth for point at Mino time τ whose image appears at screen coordinate (α, β).

Arguments

pix : Pixel information

θs : Emission Inclination

rs :Emission radius

τ : Mino Time

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

Krang.ConeGeometry — Type

struct ConeGeometry{T, A} <: Krang.AbstractGeometry

Cone Geometry with half opening angle opening_angle.

Krang.Iϕ — Function

Iϕ(pix::Krang.AbstractPixel, rs, τ, νr) -> Any

Returns the antiderivative $I_ϕ=\int\frac{a(2Mr-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

pix: SlowLightIntensityPixel

rs : Emission radius

τ : Mino time

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

Krang.SlowLightIntensityCamera — Type

struct SlowLightIntensityCamera{T} <: Krang.AbstractCamera

Observer sitting at radial infinity. The frame of this observer is alligned with the Boyer-Lindquist frame.

Krang.Iϕ_w_I0_terms — Function

Iϕ_w_I0_terms(metric::Kerr{T}, rs, τ, roots, νr, λ) -> Any

Returns the antiderivative $I_ϕ=\int\frac{a(2Mr-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ with full I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

τ: Minotime

roots : Radial roots

λ : Reduced azimuthal angular momentum

Krang.SlowLightIntensityPixel — Type

struct SlowLightIntensityPixel{T} <: Krang.AbstractPixel

Intensity Pixel Type.

Krang.radial_m_I0_terms_integrals_case4 — Function

radial_m_I0_terms_integrals_case4(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T
) -> Tuple

Returns the radial integrals for the case where there are no real roots in the radial potential

Krang.Iϕ_inf — Function

Iϕ_inf(metric::Kerr{T}, roots, λ) -> Any

Returns the antiderivative $I_ϕ=\int\frac{a(2Mr-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ evaluated at infinity without I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

roots : Radial roots

λ : Reduced azimuthal angular momentum

Krang.radial_m_I0_terms_integrals_case2 — Function

radial_m_I0_terms_integrals_case2(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    νr
) -> Tuple

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Krang.radial_inf_integrals_case3 — Function

radial_inf_integrals_case3(
    metric::Kerr{T},
    roots::NTuple{4, T} where T
) -> NTuple{4, Any}

Returns the radial integrals for the case where there are two real roots in the radial potential

Krang.AbstractPixel — Type

abstract type AbstractPixel

Abstract Pixel Type

Krang.radial_w_I0_terms_integrals_case2 — Function

radial_w_I0_terms_integrals_case2(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ,
    νr
) -> Tuple

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Krang.SlowLightIntensityScreen — Type

struct SlowLightIntensityScreen{T} <: Krang.AbstractScreen

Screen made of Intensity Pixels.

Krang.radial_w_I0_terms_integrals_case4 — Function

radial_w_I0_terms_integrals_case4(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> Tuple

Returns the radial integrals for the case where there are no real roots in the radial potential

Krang.radial_integrals — Function

radial_integrals(
    pix::Krang.AbstractPixel,
    rs,
    τ,
    νr
) -> NTuple{5, Any}

Return the radial integrals

pix: SlowLightIntensityPixel

rs : Emission radius

τ : Mino time

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

Krang.AbstractCamera — Type

abstract type AbstractCamera

Abstract Observer Type

Krang.IntensityScreen — Type

struct IntensityScreen{T} <: Krang.AbstractScreen

Screen made of Intensity Pixels.

Krang.It — Function

It(pix::Krang.AbstractPixel, rs, τ, νr) -> Any

Returns the antiderivative $I_t=\int\frac{a(2Mr-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

pix: SlowLightIntensityPixel

rs : Emission radius

τ : Mino time

νr : Sign of radial velocity direction at emission. This is always positive for case 3 and case 4 geodesics.

Krang.IntensityPixel — Type

struct IntensityPixel{T} <: Krang.AbstractPixel

Intensity Pixel Type. Each Pixel is associated with a single ray, and caches some information about the ray.

Krang.Iϕ_m_I0_terms — Function

Iϕ_m_I0_terms(metric::Kerr{T}, rs, roots, νr, λ) -> Any

Returns the antiderivative $I_ϕ=\int\frac{a(2Mr-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ with I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

τ: Minotime

roots : Radial roots

λ : Reduced azimuthal angular momentum

Krang.AbstractMaterial — Type

abstract type AbstractMaterial

Abstract Material

Krang.AbstractScreen — Type

abstract type AbstractScreen

Abstract Screen Type

Krang.radial_w_I0_terms_integrals_case3 — Function

radial_w_I0_terms_integrals_case3(
    metric::Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> Tuple

Returns the radial integrals for the case where there are two real roots in the radial potential

Krang.IntensityCamera — Type

struct IntensityCamera{T} <: Krang.AbstractCamera

Observer sitting at radial infinity. The frame of this observer is alligned with the Boyer-Lindquist frame.

Krang.Ir_s — Function

Ir_s(metric::Kerr{T}, rs, roots, νr) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$. See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

rs : Emission radius

roots : Roots of the radial potential

Krang.AbstractGeometry — Type

abstract type AbstractGeometry

Abstract Geometry

Krang.radial_inf_integrals_case2 — Function

radial_inf_integrals_case2(
    metric::Kerr{T},
    roots::NTuple{4, T} where T
) -> NTuple{4, Any}

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Krang.It_inf — Function

It_inf(
    metric::Kerr{T},
    roots::NTuple{4, T} where T,
    λ
) -> Any

Returns the antiderivative $I_t=\int\frac{r^2\Delta+2Mr(r^2+a^2-a\lambda)}{\sqrt{\Delta\mathcal{R(r)}}}dr$ evaluated at infinity without I0 terms.

See r_potential(x) for an implementation of $\mathcal{R}(r)$.

Arguments

metric: Kerr{T} metric

roots : Radial roots

λ : Reduced azimuthal angular momentum