Kang api

Raytracing Functions

Kang.AbstractMetric — Type

abstract type AbstractMetric

Abstract Metric Type

Kang.Kerr — Type

struct Kerr{T} <: Kang.AbstractMetric

Kerr Metric in Boyer Lindquist Coordinates

mass: M = mass

spin: a = J/M, where J is the angular momentum and M is the mass of the blackhole.

Kang.Gθ — Method

Gθ(
    metric::Kang.Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> NTuple{5, Any}

Returns the antiderivative $G_\theta=\int\frac{d\theta}{\sqrt{\Theta(\theta)}}$

Arguments

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image ordered by minotime

Kang.Ir — Method

Ir(metric::Kang.Kerr, νr::Bool, θo, rs, α, β) -> Any

Returns the antiderivative $I_r=\int\frac{dr}{\sqrt{\mathcal{R(r)}}}$

Arguments

metric: Kerr{T} metric

νr : Sign of radial velocity direction at emission

θo : Observer inclination

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

Kang._isreal2 — Method

_isreal2(num::Complex{T}) -> Any

Checks if a complex number is real to some tolerance

Kang.emission_coordinates — Method

emission_coordinates(
    metric::Kang.Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> NTuple{6, Any}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.emission_coordinates_fast_light_inclination_parameterized — Method

emission_coordinates_fast_light_inclination_parameterized(
    metric::Kang.Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> Tuple{Any, Any, Any, Bool, Bool}

Emission radius and azimuthal angle for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). for an observer located at inclination θo.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θs : Emission Inclination

θo : Observer Inclination

isindir : Whether emission to observer is direct or indirect

n : Image index

Kang.emission_inclination_mino_time_parameterized — Method

emission_inclination_mino_time_parameterized(
    metric::Kang.Kerr{T},
    α,
    β,
    θo,
    τ
) -> NTuple{6, Any}

Emission inclination for point at Mino time τ whose image appears at screen coordinate (α, β).

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.emission_inclination_radius_parameterized — Method

emission_inclination_radius_parameterized(
    metric::Kang.Kerr{T},
    α,
    β,
    θo,
    rs,
    νr
) -> NTuple{6, Any}

Emission inclination for point originating at inclination rs whose nth order image appears at screen coordinate (α, β).

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.emission_radius_inclination_parameterized — Method

emission_radius_inclination_parameterized(
    metric::Kang.Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> Tuple{Any, Bool, Bool}

Emission radius for point originating at inclination θs whose nth order image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

θs : Emission inclination

θo : Observer inclination

isindir : Is emission to observer direct or indirect

n : Image index

Kang.emission_radius_mino_time_parameterized — Method

emission_radius_mino_time_parameterized(
    metric::Kang.Kerr{T},
    α,
    β,
    τ,
    θo
) -> Tuple{Any, Any, Any}

Emission radius for point originating at at Mino time τ whose image appears at the screen coordinate (α, β). Returns NaN if the emission coordinates do not exist for that screen coordinate.

Arguments

-metric : Kerr{T} metric -α : Horizontal Bardeen screen coordinate -β : Vertical Bardeen screen coordinate -θs : Emission inclination -θo : Observer inclination -isindir : Normalized angular momentum -n : Image index

Kang.get_radial_roots — Method

get_radial_roots(
    metric::Kang.Kerr{T},
    η,
    λ
) -> NTuple{4, Any}

Returns roots of $r^4 + (a^2-η-λ^2)r^2 + 2(η+(a-λ)^2)r - a^2η$

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal agular momentum

Kang.horizon — Method

horizon(metric::Kang.Kerr{T}) -> Any

Outer Horizon for the Kerr metric.

Kang.metric_dd — Method

metric_dd(metric::Kang.AbstractMetric, args...) -> Any

Returns the metric in some representation (usually as an nxn matrix).

Kang.metric_dd — Method

metric_dd(metric::Kang.Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

Kang.metric_dd — Method

metric_dd(metric::Kang.Kerr{T}, coordinates) -> Any

Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Kang.metric_uu — Method

metric_uu(metric::Kang.AbstractMetric, args...) -> Any

Returns the inverse metric in some representation (usually as an nxn matrix).

Kang.metric_uu — Method

metric_uu(metric::Kang.Kerr{T}, r, θ) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

Kang.metric_uu — Method

metric_uu(metric::Kang.Kerr{T}, coordinates) -> Any

Inverse Kerr metric in Boyer Lindquist (BL) coordinates.

Arguments

metric : Kerr metric

coordinates : Coordinates (t, r, θ, ϕ)

Kang.mino_time — Method

mino_time(
    metric::Kang.Kerr{T},
    α,
    β,
    θs,
    θo,
    isindir,
    n
) -> Any

Mino time of trajectory between two inclinations for a given screen coordinate

Arguments

α : Horizontal Bardeen screen coordinate

β : Vertical Bardeen screen coordinate

a : Blackhole angular Momentum

θs : Emission inclination

θo : Observer inclination

isindir : Is the path direct or indirect?

n : nth image in orde of amount of minotime traversed

Kang.r_potential — Method

r_potential(metric::Kang.Kerr{T}, η, λ, r) -> Any

Radial potential of spacetime

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal agular momentum

r : Boyer Lindquist radius

Kang.radial_integrals_case2 — Method

radial_integrals_case2(
    metric::Kang.Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ,
    νr
) -> NTuple{5, Any}

Returns the radial integrals for the case where there are four real roots in the radial potential, with roots outside the horizon.

Kang.radial_integrals_case3 — Method

radial_integrals_case3(
    metric::Kang.Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> NTuple{5, Any}

Returns the radial integrals for the case where there are two real roots in the radial potential

Kang.radial_integrals_case4 — Method

radial_integrals_case4(
    metric::Kang.Kerr{T},
    rs,
    roots::NTuple{4, T} where T,
    τ
) -> Tuple

Returns the radial integrals for the case where there are no real roots in the radial potential

Kang.raytrace — Method

raytrace(
    metric::Kang.Kerr{T},
    α,
    β,
    θo,
    τ
) -> NTuple{6, Any}

Raytrace a point that appears at the screen coordinate (α, β) for an observer located at inclination θo

Arguments

metric : Kerr{T} metric

α : Horizontal Bardeen Screen Coordinates

β : Vertical Bardeen Screen Coordinates

θo : Observer Inclination

τ : Mino Time

Kang.regularized_Pi — Method

regularized_Pi(n, ϕ, k) -> Any

Regularized elliptic integral of the third kind

Arguments

n: Parameter

ϕ: Arguments

k: Parameter

Kang.α — Method

α(_::Kang.Kerr, λ, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination

Kang.αboundary — Method

αboundary(metric::Kang.Kerr, θs) -> Any

Defines a horizontal boundary on the assmyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr metric

θs : Emission Inclination

Kang.β — Method

β(metric::Kang.Kerr, λ, η, θo) -> Any

Horizontal Bardeen Screen Coordinate

Arguments

metric: Kerr

λ: Energy reduced Azimuthal angular momentul

η: Energy reduced Carter integral

θo: Observer inclination

Kang.βboundary — Method

βboundary(metric::Kang.Kerr{T}, α, θo, θs) -> Any

Defines a vertical boundary on the Assyptotic observers screen where emission that originates from θs must fall within.

Arguments

metric: Kerr{T} metric

α : Horizontal Bardeen screen coordinate

θo : Observer inclination

θs : Emission Inclination

Kang.η — Method

η(metric::Kang.Kerr, α, β, θo) -> Any

Energy reduced Carter integral

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

β: Bardeen vertical coordinate

θo: Observer inclination

Kang.θ_potential — Method

θ_potential(metric::Kang.Kerr{T}, η, λ, θ) -> Any

Theta potential of a Kerr blackhole

Arguments

metric: Kerr{T} metric

η : Reduced Carter constant

λ : Reduced azimuthal agular momentum

θ : Boyer Lindquist inclination

Kang.λ — Method

λ(_::Kang.Kerr, α, θo) -> Any

Energy reduced azimuthal angular momentum

Arguments

metric: Kerr

α: Horizontal Bardeen screen coordinate

θo: Observer inclination